+++ packages/runtime/src/index.ts
@@ -6,3 +6,4 @@ export * from "./valuedashboard"
 export * from "./map"
 export * from "./set"
 export * from "./pixelbuffer"
+export * from "./led"
